Pontos refactor

+++ Trabalho_02/v1/auxiliares.h
@@ -13,11 +13,11 @@
 
 // Parâmetros: número inteiro
 // Retorno: endereço do vetor alocado de tamanho N
-struct operandos* alocarVetor(int N);
+struct operandos* alocarVetor(long long int N);
 
 // Parâmetros: número inteiro
 // Retorno: endereço da matriz alocada de tamanho NxN
-struct operandos** alocarMatriz(int N);
+struct operandos** alocarMatriz(long long int N);
 
 // Parâmetro: vetor de double [4]
 // Retorno: maior número do vetor (double)
